==> notebooks/1.0-ditp-seoul-bicycle-traffic-data-downloader.ipynb <==
pip install openml
import openml
import pandas as pd

# Use if working in colab
#from google.colab import drive
#drive.mount('/content/drive')

# Use if working in colab
# set working directory to project directory
#%cd '/content/drive/My Drive/github/bicycles-seoul'

# Get dataset by id
dataset = openml.datasets.get_dataset(46328)

# Get the data itself as a dataframe (or otherwise)
X, y, _, _ = dataset.get_data(dataset_format="dataframe")
# merge into a single df
df = X.copy()
# in this case, y is empty. no need to merge with code below
#df['target'] = y

# Save as csv in Drive
df.to_csv('data/raw/bicycles_seoul_raw.csv', index=False)
