==> notebooks/1.1-ditp-seoul-pollution-data-downloader.ipynb <==
# Use if working in colab
from google.colab import drive
drive.mount('/content/drive')

# Use if working in colab
# set working directory to project directory
%cd '/content/drive/My Drive/github/bicycles-seoul'

import os
import pandas as pd
import kagglehub

# Download the dataset - returns a directory that contains the extracted files
dir_path = kagglehub.dataset_download("bappekim/air-pollution-in-seoul")

print("Directory contents:", os.listdir(dir_path))
# From your output, 'AirPollutionSeoul' is inside dir_path, and 'Measurement_summary.csv' is inside 'AirPollutionSeoul'.

measurements_file = os.path.join(dir_path, "AirPollutionSeoul", "Measurement_summary.csv")
df = pd.read_csv(measurements_file)
df.head()

# Save as csv in Drive
df.to_csv('data/external/seoul-pollution.csv', index=False)
