==> notebooks/2.0-ditp-EDA.ipynb <==
import pandas as pd
import plotly.express as px

# Use if working in colab
from google.colab import drive
drive.mount('/content/drive')

# Use if working in colab
# set working directory to project directory
%cd '/content/drive/My Drive/github/bicycles-seoul'

df = pd.read_csv('data/interim/01-merged-data.csv')

df.head()

# Efficiently check for dtypes, non-null values, unique values, NaN values.

def summarize_df(df):
    summary = pd.DataFrame({
        'Data Type': df.dtypes,
        'Non-Null Count': df.notnull().sum(),
        'Unique Values': df.nunique(),
        'NaN Count': df.isnull().sum()
    })
    return summary


summary = summarize_df(df)
print(summary)

df.describe().T

# Day, month, year, hour, and weekday have already been extracted from timestamp/datetime, so drop timestamp/datetime.

df = df.drop(columns=['timestamp', 'datetime'])

# Divide the variables into numeric and categoric, and analyse accordingly

# What is functioning day, and what relationship does it bear to rented bike count?

fig = px.bar(
    df.groupby("Functioning Day")["rented_bike_count"].sum().reset_index(),
    x="Functioning Day",
    y="rented_bike_count",
    title="Total Rented Bike Count by Functioning Day",
    labels={"rented_bike_count": "Total Bike Count", "Functioning Day": "Functioning Day"}
)

fig.show()

# Functioning Day determines whether the bike service is running - so needs to be removed.
# Drop all cases where Functioning Day = No. Then delete Functioning Day variable.

print('shape before modification', df.shape)
df = df[df["Functioning Day"] != "No"]
df = df.drop(columns=["Functioning Day"])
print('shape after modification', df.shape)

# Save as csv in Drive
df.to_csv('data/interim/02-data-after-EDA.csv', index=False)
